경고 부분 수정
Project : Engine_SOURCE

[Modify]

M1. hjCamera.cpp
M2. hjConstantBuffer.cpp
M3. hjEntity.cpp
M4. hjInput.cpp
M5. hjTime.cpp

+++ Engine_SOURCE/hjCamera.cpp
@@ -69,8 +69,8 @@ namespace hj
 		RECT winRect;
 		GetClientRect(application.GetHwnd(), &winRect);
 
-		float width = winRect.right - winRect.left;
-		float height = winRect.bottom - winRect.top;
+		float width = static_cast<float>(winRect.right - winRect.left);
+		float height = static_cast<float>(winRect.bottom - winRect.top);
 		mAspectRatio = width / height;
 
 		if (mType == eProjectionType::Perspective)
