+++ Engine_SOURCE/hjConstantBuffer.cpp
@@ -13,7 +13,7 @@ namespace hj::graphics
 	bool ConstantBuffer::Create(size_t size)
 	{
 		// 상수 버퍼 명세		
-		desc.ByteWidth = size;
+		desc.ByteWidth = static_cast<UINT>(size);
 		desc.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_CONSTANT_BUFFER;
 		desc.Usage = D3D11_USAGE::D3D11_USAGE_DYNAMIC;
 		desc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
