+++ Engine_SOURCE/hjTime.cpp
@@ -46,7 +46,7 @@ namespace hj
             wchar_t szFloat[50] = {};
             float FPS = 1.f / mDeltaTime;
             swprintf_s(szFloat, 50, L"DeltaTime : %d", iCount);
-            int iLen = wcsnlen_s(szFloat, 50);
+            int iLen = static_cast<int>(wcsnlen_s(szFloat, 50));
             //TextOut(_dc, 10, 10, szFloat, iLen);
 
             SetWindowText(hWnd, szFloat);
